refactor: add sortVotesByType util
