+++ src/proposals/proposals.service.ts
@@ -8,8 +8,8 @@ import { ImagesService } from "../images/images.service";
 import { Image } from "../images/models/image.model";
 import { User } from "../users/models/user.model";
 import { Vote } from "../votes/models/vote.model";
-import { VoteTypes } from "../votes/votes.constants";
 import { VotesService } from "../votes/votes.service";
+import { sortVotesByType } from "../votes/votes.utils";
 import { CreateProposalInput } from "./models/create-proposal.input";
 import { Proposal } from "./models/proposal.model";
 import {
@@ -17,13 +17,6 @@ import {
   MIN_VOTE_COUNT_TO_RATIFY,
   ProposalStages,
 } from "./proposals.constants";
-
-interface SortedVotes {
-  agreements: Vote[];
-  reservations: Vote[];
-  standAsides: Vote[];
-  blocks: Vote[];
-}
 
 @Injectable()
 export class ProposalsService {
@@ -125,29 +118,7 @@ export class ProposalsService {
       DefaultGroupSettings.RatificationThreshold * 0.01;
 
     const { agreements, reservations, standAsides, blocks } =
-      votes.reduce<SortedVotes>(
-        (result, vote) => {
-          if (vote.voteType === VoteTypes.Agreement) {
-            result.agreements.push(vote);
-          }
-          if (vote.voteType === VoteTypes.Reservations) {
-            result.reservations.push(vote);
-          }
-          if (vote.voteType === VoteTypes.StandAside) {
-            result.standAsides.push(vote);
-          }
-          if (vote.voteType === VoteTypes.Block) {
-            result.blocks.push(vote);
-          }
-          return result;
-        },
-        {
-          agreements: [],
-          reservations: [],
-          standAsides: [],
-          blocks: [],
-        }
-      );
+      sortVotesByType(votes);
 
     return (
       agreements.length >= members.length * ratificationThreshold &&
